pyproject.toml
[project]
name = "cv-builder"
version = "0.1.0"
description = "Multi-agent system for generating personalized CVs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "litellm~=1.67.0",
    "pydantic~=2.10.0",
    "pyyaml~=6.0.2",
    "python-dotenv~=1.0.1",
    "httpx~=0.28.1",
    "boto3~=1.36.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.1",
]

[project.optional-dependencies]
dev = [
    "pytest~=8.3.0",
    "pytest-asyncio~=0.25.0",
]

[project.scripts]
cv-builder = "src.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
addopts = "-m 'not integration'"
markers = [
    "integration: tests that require real .env config and LLM/Google Drive connectivity",
    "codex_integration: tests that require the local Codex bridge and authentication",
]
